requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
numpy>=1.26.4
requests>=2.31.0
openai>=1.30.0
plotly>=5.22.0
python-dotenv>=1.0.1
yfinance>=0.2.40
scipy>=1.13.0
